Matlab code reads Excel sheet draw complexity
In number theory, the integer complexity of an integer is the smallest number of ones that can be used to represent it using ones and any number of additions, multiplications, and parentheses. It is always within a constant factor of the logarithm of the given integer. The property is used to charaterize the hacking bound of the multivariate polynomials as an example, it applies to lattice as well. Graph and randomly generated permutation of the edge set (1, . .  , n) is used to create the isomorphic graph for cryptography key pairs.